requirements.txt
flask
pyTelegramBotAPI
requests
django
waitress
urlopen
beautifulsoup4
lxml
sports.py